+++ src/winids.h
@@ -28,6 +28,7 @@
 #define IDM_TEKPAGE         0x0260
 #define IDM_TEKCOPY         0x0360
 #define IDM_SAVEIMG         0x0460
+#define IDM_RESET_NOASK     0x0560
 #define IDM_DEFSIZE         0x0070
 #define IDM_DEFSIZE_ZOOM    0x0170
 #define IDM_SCROLLBAR       0x0280